requirements2.txt
g2p-en == 2.1.0
inflect == 4.1.0
librosa == 0.8.0
matplotlib == 3.7.1
numba == 0.57.0
numpy == 1.23.5
pypinyin==0.39.0
pyworld == 0.3.3
PyYAML==5.4.1
scikit-learn==1.2.2
scipy == 1.10.1
soundfile==0.10.3.post1
tensorboard == 2.2.2
tgt == 1.4.4
torch == 2.0.0
tqdm==4.46.1
unidecode == 1.1.1